+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@
 * Added single-stack IPv6 and dual-stack networking support for Kubernetes
 * SUSEConnect now properly activates the SL Micro "Extras" module
 * Improved Embedded Artifact Registry handling to no longer be memory bound
+* Improved Helm chart handling to allow deploying multiple Helm charts with the same chart name
 * Dependency upgrades
   * Updated metallb from `0.14.9` to `0.1.0+up0.14.9`
 
@@ -25,6 +26,7 @@
 * [#593](https://github.com/suse-edge/edge-image-builder/issues/593) - OS files script should mount /var
 * [#594](https://github.com/suse-edge/edge-image-builder/issues/594) - Package installation breaks package resolution if packages are already installed on root OS
 * [#632](https://github.com/suse-edge/edge-image-builder/issues/632) - Create the required Elemental Agent directory structure during Combustion
+* [#632](https://github.com/suse-edge/edge-image-builder/issues/606) - Allow for duplicate Helm chart names
 
 ---